+++ src/main/java/shop/tryit/domain/order/OrderDetailRepository.java
@@ -3,6 +3,7 @@ package shop.tryit.domain.order;
 import java.util.List;
 import java.util.Optional;
 import shop.tryit.domain.order.dto.OrderDetailSearchDto;
+import shop.tryit.domain.order.entity.Order;
 
 public interface OrderDetailRepository {
