+++ src/main/java/shop/tryit/repository/order/OrderDetailCustom.java
@@ -1,8 +1,8 @@
 package shop.tryit.repository.order;
 
 import java.util.List;
-import shop.tryit.domain.order.Order;
 import shop.tryit.domain.order.dto.OrderDetailSearchDto;
+import shop.tryit.domain.order.entity.Order;
 
 public interface OrderDetailCustom {
