+++ src/main/java/shop/tryit/repository/order/OrderJpaRepository.java
@@ -1,7 +1,7 @@
 package shop.tryit.repository.order;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.tryit.domain.order.Order;
+import shop.tryit.domain.order.entity.Order;
 
 public interface OrderJpaRepository extends JpaRepository<Order, Long>, OrderCustom {
