+++ src/main/java/shop/tryit/repository/order/OrderRepositoryImpl.java
@@ -7,9 +7,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 import shop.tryit.domain.member.Member;
-import shop.tryit.domain.order.Order;
 import shop.tryit.domain.order.OrderRepository;
 import shop.tryit.domain.order.dto.OrderSearchDto;
+import shop.tryit.domain.order.entity.Order;
 
 @Slf4j
 @Repository
